requirements.txt
fastapi
uvicorn
langchain-groq
langchain-core
mediapipe
box2d-py
python-multipart
gtts
python-dotenv
pydantic
groq
pinecone
sentence-transformers